backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.42
psycopg[binary]==3.2.13
pydantic==2.11.7
pydantic-settings==2.10.1
PyJWT==2.10.1
email-validator==2.3.0
